User service unit tests

+++ src/main/java/com/salarytracker/salaryapp/repository/UserRepositoryCustom.java
@@ -3,5 +3,5 @@ package com.salarytracker.salaryapp.repository;
 import java.util.List;
 
 public interface UserRepositoryCustom {
-    List<User> findBySalary(Long min, Long max);
+    List<User> findBySalary(Double min, Double max);
 }
